Move db_session and test_client fixtures into conftest.py
